dreamdiffusion/requirements.txt
absl-py==2.1.0
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.12
altair==5.0.1
antlr4-python3-runtime==4.8
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.14.2
astropy==5.1
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==23.1.0
Automat==20.2.0
autopep8==2.0.4
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.4.2
bleach==4.1.0
blinker==1.6.2
bokeh==2.4.3
botocore==1.34.69
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.1
constantly==23.10.4
contourpy==1.0.5
cookiecutter==2.6.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.11.0
cytoolz==0.12.2
dask==2023.4.1
dataclasses==0.6
datashader==0.15.1
datashape==0.5.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.8
distributed==2023.4.1
docker-pycreds==0.4.0
docstring-to-markdown==0.11
docutils==0.18.1
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
flake8==7.0.0
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2024.3.1
ftfy==6.2.3
future==1.0.0
gensim==4.3.2
gitdb==4.0.7
GitPython==3.1.37
gmpy2==2.1.2
google-auth==2.33.0
google-auth-oauthlib==1.0.0
greenlet==3.0.1
grpcio==1.65.4
h5py==3.11.0
HeapDict==1.0.1
holoviews==1.17.1
huggingface-hub==0.24.5
hvplot==0.10.0
hyperlink==21.0.0
idna==3.7
imagecodecs==2023.1.23
imageio==2.33.1
imagesize==1.4.1
imbalanced-learn==0.12.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.7.0
intervaltree==3.1.0
ipykernel==6.28.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==7.8.1
isort==5.13.2
itemadapter==0.3.0
itemloaders==1.1.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==1.0.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==1.0.0
keyring==24.3.1
kiwisolver==1.4.4
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lazy-object-proxy==1.10.0
lckr_jupyterlab_variableinspector==3.1.0
lightning-utilities==0.11.6
llvmlite==0.41.0
lmdb==1.4.1
locket==1.0.0
lpips==0.1.4
lxml==5.2.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.0
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
multipledispatch==0.6.0
mypy==1.10.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.4
numpy==1.24.3
numpydoc==1.5.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
panel==0.14.3
param==1.13.0
parsel==1.8.1
parso==0.8.3
partd==1.4.1
pathspec==0.10.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.22.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
Protego==0.1.16
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycparser==2.21
pyct==0.5.0
pycurl==7.45.2
pydeck==0.8.0
pyDeprecate==0.3.2
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyerfa==2.0.0
pyflakes==3.2.0
Pygments==2.15.1
pylint==2.16.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
pynndescent==0.5.13
pyodbc==5.0.1
pyOpenSSL==24.0.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-slugify==5.0.2
python-snappy==0.6.1
pytoolconfig==1.2.6
pytorch-lightning==1.6.5
pytorch-metric-learning==2.6.0
pytz==2024.1
pyviz_comms==3.0.2
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.10.3
requests==2.32.2
requests-file==1.5.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rope==1.12.0
rpds-py==0.10.6
rsa==4.9
Rtree==1.0.1
s3fs==2024.3.1
safetensors==0.4.4
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.9.1
Scrapy==2.11.1
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==2.13.0
service-identity==18.1.0
setproctitle==1.3.3
setuptools==69.5.1
sip==6.7.12
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.5.1
spyder-kernels==2.5.0
SQLAlchemy==2.0.30
stack-data==0.2.0
statsmodels==0.14.0
streamlit==1.32.0
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2023.4.12
timm==0.5.4
tinycss2==1.2.1
tldextract==3.2.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==2.4.0
torchaudio==0.7.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
Twisted==23.10.0
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.10.0
umap-learn==0.5.6
unicodedata2==15.1.0
Unidecode==1.2.0
urllib3==1.26.19
w3lib==2.1.2
wandb==0.17.6
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
whatthepatch==1.0.2
wheel==0.43.0
widgetsnbextension==3.6.6
wrapt==1.14.1
wurlitzer==3.0.2
xarray==2022.11.0
yapf==0.40.2
yarl==1.9.3
zict==3.0.0
zipp==3.17.0
zope.interface==5.4.0
